move recovery related logic to its own file
